% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conf_metrics.R
\name{get_conf_metrics}
\alias{get_conf_metrics}
\title{Calculate diagnostics}
\usage{
get_conf_metrics(pred_df, truth_vec = NULL, pred_vec = NULL)
}
\arguments{
\item{pred_df}{Dataframe of truth and predicted classes for each case.}

\item{truth_vec}{Vector of true classes.}

\item{pred_vec}{Vector of predicted classes.}
}
\value{
Tibble of metrics.
}
\description{
Takes vectors of truth and predicted cases and returns a range of \code{yardstick}
metrics.
}
