% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappa}
\alias{kappa}
\title{Calculate kappa statistics from confusion matrix}
\usage{
kappa(CM, by_class)
}
\arguments{
\item{CM}{Confusion matrix.}

\item{by_class}{Logical. Return a row per class, or mean values across all
classes.}
}
\value{
Tibble
}
\description{
Code from \emph{Technical Note: Statistical methods for accuracy assessment of
classified thematic maps} \insertRef{rossiter2004statistical}{envModel}.
}
