% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_env_cluster_df.R
\name{make_env_clust_df}
\alias{make_env_clust_df}
\title{Add explanatory (environmental) variables to 'site' dataframe.}
\usage{
make_env_clust_df(
  df,
  clustcol = "cluster",
  sitecol = "cell",
  minsites = minAbsSites,
  additionalClusters,
  dfenv,
  addenv,
  setmin = FALSE,
  envcols = names(dfEnv)[-1]
)
}
\arguments{
\item{df}{Dataframe with 'site' and 'cluster' columns.}

\item{clustcol}{Name of cluster column.}

\item{sitecol}{Name of 'site' column.}

\item{minsites}{Minimum number of sites prefered in a cluster.}

\item{additionalClusters}{Any additional clusters (i.e. clusters not
generated by floristics such as land cover 'clusters')}

\item{dfenv}{Dataframe with 'site' and associated env values for floristic
'sites'.}

\item{addenv}{Dataframe with 'site' and associated env values for
additionalClusters 'sites'.}

\item{setmin}{Filter clusters those with 'sites' >= minsites}

\item{envcols}{Names of the columns containing the env variables.}
}
\value{
Dataframe of siteCol, clustCol and envCols from both dfenv and addenv.
}
\description{
Add explanatory (environmental) variables to 'site' dataframe.
}
