% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_env_corr.R
\name{make_env_corr}
\alias{make_env_corr}
\title{Generate correlation matrix and select variables to remove}
\usage{
make_env_corr(
  env_df,
  env_cols,
  remove = TRUE,
  thresh = 0.95,
  always_remove = c("lat", "long"),
  always_keep = c("prec", "temp")
)
}
\arguments{
\item{env_df}{Dataframe with environmental variables}

\item{env_cols}{Numeric or character vector defining columns in env_df
that contain the variables to test for correlation.}

\item{remove}{Logical. Remove variables correlated at or above \code{thresh}.
Correlation is found via \code{\link[caret:findCorrelation]{caret::findCorrelation()}}.}

\item{thresh}{Numeric. Threshold correlation value before variable is removed.}

\item{always_remove}{Character vector. Any string in this vector, if matched,
will always be removed (even if not correlated).}

\item{always_keep}{Character vector. Any string in this vector, if matched,
will not be removed (even if correlated).}
}
\value{
list with elements
itemize{
\item{env_cols}{as provided}
\item{always_remove}{as provided}
\item{always_keep}{as provided}
\item{remove}{as provided}
\item{env_corr}{result of applying \code{\link[stats:cor]{stats::cor()}} to \code{env_df[,env_cols]}}
\item{highly_corr}{column names in \code{env_df} that are correlated with another
column above \code{thresh}.}
\item{remove_env}{column names to remove, based on inputs}
}
}
\description{
Generate correlation matrix and select variables to remove
}
