% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conf_metrics.R
\name{make_model_diagnostics}
\alias{make_model_diagnostics}
\title{Calculate diagnostics}
\usage{
make_model_diagnostics(df_pred, by_class)
}
\arguments{
\item{df_pred}{Data frame containing truth and predicted classes.}

\item{by_class}{Logical. Return a row per class, or mean values across all
classes.}
}
\value{
Tibble
}
\description{
Takes a data frame of truth and predicted cases. Returns overall or per
class results.
}
