% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rf_diagnostics_spatialcv.R
\name{make_rf_diagnostics_spatialcv}
\alias{make_rf_diagnostics_spatialcv}
\title{Run random forest, returning only diagnostic values from spatial cross
validation.}
\usage{
make_rf_diagnostics_spatialcv(
  env_df,
  clust_col = "cluster",
  folds = 3L,
  reps = 5L,
  range_min = 20000,
  range_max = 1e+05,
  coords = c("long", "lat"),
  crs_df = 4283,
  crs_analysis = 3577,
  ...
)
}
\arguments{
\item{env_df}{Dataframe with clusters and environmental columns.}

\item{clust_col}{Character. Name of column with cluster membership.}

\item{folds}{Numeric. How many folds to use in cross-validation?}

\item{reps}{Numeric. How many repeats of cross-validation?}

\item{range_min, range_max}{Numeric. Metres. Minimum and maximum grid sizes
for spatial tiles. Reps will be made using \code{range <- seq(range_min, range_max , length.out = reps)}}

\item{coords}{Character vector of length 2. Names of columns in \code{env_df} with
x and y coordinates.}

\item{crs_df}{Coordinate reference system for \code{coords}. Passed to the
\code{crs} argument of \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}

\item{...}{passed to \code{\link[=make_rf_good]{make_rf_good()}}.}

\item{crs_analysis.}{Coordinate reference system for the analysis. Should be
a projected crs so that \code{range_min} and \code{range_max} are interpreted
correctly.}

\item{set_min}{FALSE or numeric. If numeric, classes in \code{clust_col} with less
than \code{set_min} cases will be filtered.}
}
\value{

}
\description{
Random forest is run via \code{\link[randomForest:randomForest]{randomForest::randomForest()}}.
}
