% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{make_rf_good}
\alias{make_rf_good}
\title{Iteratively add trees to random forest until predictions stabilise}
\usage{
make_rf_good(
  env_df,
  clust_col = "cluster",
  context = "cell",
  env_names,
  trees_start = 499,
  trees_add = 249,
  trees_max = 9999,
  rf_cores = 1,
  use_mtry = NULL,
  out_file = NULL,
  accept_prev_delta = 0.995,
  accept_prev_kappa = 0.995
)
}
\arguments{
\item{env_df}{Dataframe with \code{clust_col}, \code{site_col} and columns \code{env_names}}

\item{clust_col}{Character. Name of the columns with clusters.}

\item{context}{Character. Name of the columns defining the context.}

\item{env_names}{Character. Name of the environmental variables (e.g.
\code{names(stack_list)}).}

\item{trees_start}{Number of trees in first random forest run.}

\item{trees_add}{Number of trees to add in each subsequent run.}

\item{rf_cores}{Number of cores to use for parallel processing.}

\item{use_mtry}{\code{mtry} value for \code{randomForest} call. If \code{NULL} it will be
generated by a (lengthy) call to \code{caret::train} with a tune grid of
\verb{.mtry = 1:floor(sqrt(length(env_names))}}

\item{out_file}{Optional name of file to save results.}
}
\value{

}
\description{
Iteratively add trees to random forest until predictions stabilise
}
