% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_env.R
\name{reduce_env}
\alias{reduce_env}
\title{Reduce number of environmental variables}
\usage{
reduce_env(
  env_df,
  env_cols,
  y_col = NULL,
  remove = TRUE,
  thresh = 0.95,
  remove_always = c("lat", "long"),
  keep_always = c("rain", "tavg")
)
}
\arguments{
\item{env_df}{Dataframe with environmental variables}

\item{env_cols}{Numeric or character vector defining columns in env_df
that contain the variables to test for correlation. Correlation is found via
\code{\link[caret:findCorrelation]{caret::findCorrelation()}}.}

\item{y_col}{Character. Name of a column in \code{env_df} to test for
'important' variables.}

\item{thresh}{Numeric. Threshold correlation value above which a variable
is flagged as correlated.}

\item{remove_always}{Character. Any matches will always be removed (even if
not correlated).}

\item{keep_always}{Character vector. Any string in this vector, if matched,
will not be removed (even if correlated).}

\item{y_col_factor}{Logical. Should \code{y_col} be considered a
factor in \code{randomForest::randomForest()}? Thus classification rather than
regression.}
}
\value{
list with elements
itemize{
\item{env_cols}{as provided}
\item{remove_always}{as provided}
\item{keep_always}{as provided}
\item{remove_constant}{columns with only one value}
\item{env_corr}{result of applying \code{\link[stats:cor]{stats::cor()}} to \code{env_df[,env_cols]}}
\item{remove_corr}{column names in \code{env_df} that are correlated with another
column above \code{thresh}.}
\item{rf}{result of applying \code{randomforest::randomforest()} in the form of
y_col ~ env_cols}
\item{rf_vi}{variable importance from rf}
\item{remove}{character vector of columns to remove from further
analysis, based on results from \code{reduce_env()}}
}
}
\description{
Given a data frame of environmental variables, flag correlated, and
optionally, important variables as precursor to further analysis.
}
