% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_env_cluster_df.R
\name{make_env_clust_df}
\alias{make_env_clust_df}
\title{Add explanatory (environmental) variables to 'site' dataframe.}
\usage{
make_env_clust_df(
  df,
  clust_col = "cluster",
  context = "cell",
  min_sites = min_abs_sites,
  df_env,
  add_clust,
  add_env,
  env_cols,
  set_min = FALSE
)
}
\arguments{
\item{df}{Dataframe with context and 'cluster' columns.}

\item{clust_col}{Name of cluster column.}

\item{context}{Name of columns describing the context.}

\item{min_sites}{Minimum number of sites preferred in a cluster.}

\item{df_env}{Dataframe with 'site' and associated env values for floristic
'sites'.}

\item{add_clust}{Any additional clusters (i.e. clusters not generated by
floristics such as land cover 'clusters')}

\item{add_env}{Dataframe with 'site' and associated env values for
additionalClusters 'sites'.}

\item{env_cols}{Names of the columns containing the env variables.}

\item{set_min}{Filter clusters those with 'sites' >= minsites}
}
\value{
Dataframe of context columns, clustCol and envCols with rows from
both dfenv and addenv.
}
\description{
Add explanatory (environmental) variables to 'site' dataframe.
}
