% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rf_diagnostics.R
\name{make_rf_diagnostics}
\alias{make_rf_diagnostics}
\title{Run random forest, returning only diagnostic values.}
\usage{
make_rf_diagnostics(
  env_df,
  clust_col = "cluster",
  context = "cell",
  env_cols,
  folds = 3,
  reps = 2,
  ntrees = 999
)
}
\arguments{
\item{env_df}{Dataframe with clusters, context and environmental columns.}

\item{clust_col}{Character. Name of column with cluster membership.}

\item{context}{Character. Name of column(s) defining the context.}

\item{env_cols}{Character. Name of columns with environmental data.}

\item{folds}{Numeric. How many folds to use in cross-validation?}

\item{reps}{Numeric. How many repeats of cross-validation?}

\item{ntrees}{Numeric. How many trees in the random forest?}
}
\value{

}
\description{
Random forest is run via \code{parsnip} with engine \code{randomForest}.
}
