% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{make_rf_diagnostics}
\alias{make_rf_diagnostics}
\title{Run random forest, usually returning only diagnostic values.}
\usage{
make_rf_diagnostics(
  df,
  clust_col = "cluster",
  env_cols,
  trees = 999,
  folds = 3,
  reps = 5,
  tune_length = NULL,
  save_rf = NULL
)
}
\arguments{
\item{df}{Dataframe with clusters, context and environmental columns.}

\item{clust_col}{Character. Name of column with cluster membership.}

\item{env_cols}{Character. Name of columns with environmental data.}

\item{trees}{Numeric. How many trees in the forest?}

\item{folds}{Numeric. How many folds to use in repeated cross-validation.}

\item{reps}{Numeric. How many reps to use in \code{rep}eated cross-validation.}

\item{tune_length}{Numeric. Passed to \code{caret::train} argument \code{tuneLength}.}

\item{save_rf}{Character. File name for saving the \code{train} object as (.rds).}
}
\value{
Tibble (row) of diagnostics
}
\description{
Random forest is run via \code{caret::train}.
}
