% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rf_diagnostics.R
\name{make_rf_diagnostics}
\alias{make_rf_diagnostics}
\title{Run random forest, returning only diagnostic values.}
\usage{
make_rf_diagnostics(
  env_df,
  clust_col = "cluster",
  context = "cell",
  env_names,
  folds = 3,
  reps = 5,
  use_mtry = floor(sqrt(length(env_names))),
  set_min = FALSE,
  summarise_folds = TRUE,
  trees_start = 499,
  trees_add = 499,
  trees_max = 9999,
  accept_delta = formals(make_rf_good)$accept_prev_delta,
  accept_run = 3
)
}
\arguments{
\item{env_df}{Dataframe with clusters, context and environmental columns.}

\item{clust_col}{Character. Name of column with cluster membership.}

\item{context}{Character. Name of column(s) defining the context.}

\item{env_names}{Character. Name of columns with environmental data.}

\item{folds}{Numeric. How many folds to use in cross-validation?}

\item{reps}{Numeric. How many repeats of cross-validation?}

\item{use_mtry}{Numeric. \code{mtry} value used by random forest.}

\item{set_min}{FALSE or numeric. If numeric, classes in \code{clust_col} with less
than \code{set_min} cases will be filtered.}

\item{summarise_folds}{Logical. Return mean values for folds or each fold.}

\item{trees_start, trees_add, trees_max}{passed to \verb{\link[envModel]\{make_rf_good\}}}

\item{accept_delta}{Proportion. What change in predictions (as trees are
added to the forest) is acceptable?}

\item{accept_run}{How many forests (in a row) need to beat \code{accept_delta}?}
}
\value{

}
\description{
Random forest is run via \code{parsnip} with engine \code{randomForest}.
}
