% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{make_rf_diagnostics}
\alias{make_rf_diagnostics}
\title{Run a basic random forest, saving only the confusion matrix.}
\usage{
make_rf_diagnostics(
  df,
  clust_col = "cluster",
  env_cols,
  trees = 999,
  folds = 3,
  reps = 5
)
}
\arguments{
\item{df}{Dataframe with clusters, context and environmental columns.}

\item{clust_col}{Character. Name of column with cluster membership.}

\item{env_cols}{Character. Name of columns with environmental data.}

\item{trees}{Numeric. How many trees in the forest?}

\item{out_file}{Character. Path for saving the confusion matrix.}
}
\value{

}
\description{
Run a basic random forest, saving only the confusion matrix.
}
