% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rf_diagnostics.R
\name{make_rf_diagnostics}
\alias{make_rf_diagnostics}
\title{Run random forest, returning only diagnostic values.}
\usage{
make_rf_diagnostics(
  env_df,
  clust_col = "cluster",
  folds = 3,
  reps = 5,
  set_min = FALSE,
  spatial_cv = FALSE,
  x = NULL,
  y = NULL,
  ...
)
}
\arguments{
\item{env_df}{Dataframe with clusters and environmental columns.}

\item{clust_col}{Character. Name of column with cluster membership.}

\item{folds}{Numeric. How many folds to use in cross-validation?}

\item{reps}{Numeric. How many repeats of cross-validation?}

\item{set_min}{FALSE or numeric. If numeric, classes in \code{clust_col} with less
than \code{set_min} cases will be filtered.}

\item{x}{Character name of column containing x coordinates.}

\item{y}{Character name of column containing y coordinates.}

\item{...}{passed to \code{\link[=make_rf_good]{make_rf_good()}}.}

\item{`cv_func`}{Name of function to use to create folds. Needs to create a
\code{rsplit} object.}
}
\value{

}
\description{
Random forest is run via \code{\link[randomForest:randomForest]{randomForest::randomForest()}}.
}
