% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rf_diagnostics.R
\name{make_rf_diagnostics}
\alias{make_rf_diagnostics}
\title{Run random forest, returning only diagnostic values.}
\usage{
make_rf_diagnostics(
  env_df,
  clust_col = "cluster",
  folds = 3,
  reps = 5,
  set_min = FALSE,
  summarise_folds = TRUE,
  ...
)
}
\arguments{
\item{env_df}{Dataframe with clusters, context and environmental columns.}

\item{clust_col}{Character. Name of column with cluster membership.}

\item{folds}{Numeric. How many folds to use in cross-validation?}

\item{reps}{Numeric. How many repeats of cross-validation?}

\item{set_min}{FALSE or numeric. If numeric, classes in \code{clust_col} with less
than \code{set_min} cases will be filtered.}

\item{summarise_folds}{Logical. Return mean values for folds or each fold.}

\item{...}{passed to \code{\link[=make_rf_good]{make_rf_good()}}.}
}
\value{

}
\description{
Random forest is run via \code{parsnip} with engine \code{randomForest}.
}
