% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rf_diagnostics.R
\name{make_rf_diagnostics}
\alias{make_rf_diagnostics}
\title{Run random forest, returning only diagnostic values.}
\usage{
make_rf_diagnostics(
  env_df,
  clust_col = "cluster",
  folds = 3L,
  reps = 5L,
  range_m = as.integer(seq(20000L, 100000L, length.out = reps)),
  set_min = FALSE,
  mlr3_cv_method = "repeated_cv",
  coords = c("long", "lat"),
  crs_df = 4283
)
}
\arguments{
\item{env_df}{Dataframe with clusters and environmental columns.}

\item{clust_col}{Character. Name of column with cluster membership.}

\item{folds}{Numeric. How many folds to use in cross-validation?}

\item{reps}{Numeric. How many repeats of cross-validation?}

\item{range_m}{Numeric. The distance in metres (regardless of the unit
of the reference system of the input data) for block size(s) if using
\code{\link[blockCV:spatialBlock]{blockCV::spatialBlock()}}. If reps > 1, an equivalent number of range_m
values are required to ensure the folds are different between repetitions.
\code{repeated_spcv_block}.}

\item{set_min}{FALSE or numeric. If numeric, classes in \code{clust_col} with less
than \code{set_min} cases will be filtered.}

\item{mlr3_cv_method}{Method to use with \code{\link[mlr3:mlr_sugar]{mlr3::rsmp()}} (as character, e.g.
"repeated_cv" or "repeated_spcv_block".}

\item{coords}{Character vector of length 2. Names of columns in \code{env_df} with
x and y coordinates.}

\item{crs_df}{Coordinate reference system for \code{coords}. Passed to the
\code{crs} argument of \code{\link[sf:st_as_sf]{sf::st_as_sf()}}.}
}
\description{
Run random forest, returning only diagnostic values.
}
